User(MyPetsEdit, ProfileEdit)

+++ application/controllers/Profile.php
@@ -91,11 +91,14 @@ class Profile extends CI_Controller {
         $config['max_size'] = 5120;
         $config['encrypt_name'] = true;
         $this->load->library('upload', $config);
-
         if (!empty($_FILES["user_picture"]["name"])) {
             if ($this->upload->do_upload('user_picture')) {
                 $imagePath = "images/user/" . $this->upload->data("file_name");
-                unlink($userDetails->user_picture);
+                if ($userDetails->user_picture == "images/user/male.png" || $userDetails->user_picture == "images/user/female.png") {
+                    
+                } else {
+                    unlink($userDetails->user_picture);
+                }
             } else {
                 echo $this->upload->display_errors();
                 $this->session->set_flashdata("uploading_error", "Please make sure that the max size is 5MB the types may only be .jpg, .jpeg, .gif, .png");
@@ -119,21 +122,22 @@ class Profile extends CI_Controller {
 
         if ($this->Profile_model->update_user_record($data, array("user_id" => $userDetails->user_id))) {
             //SUCCESS
-            $this->SaveEventUser->trail($this->session->userdata("userid"), $userDetails->user_firstname . "Change profile picture.");
+            $this->SaveEventUser->trail($this->session->userdata("userid"), $userDetails->user_firstname . " change profile picture.");
             $this->session->set_flashdata("uploading_success", "Successfully update the image");
+            redirect(base_url() . "Profile/edit_profile");
         } else {
             
         }
         redirect(base_url() . "Profile/");
     }
 
-    public function edit_personalInfo_submit() {
+    public function edit_profile_submit() {
         $userDetails = $this->Profile_model->fetch("user", array("user_id" => $this->session->userdata("userid")))[0];
         $this->form_validation->set_rules('user_firstname', "Firstname", "required|callback__alpha_dash_space|min_length[2]");
         $this->form_validation->set_rules('user_lastname', "Lastname", "required|callback__alpha_dash_space|min_length[2]");
-        $this->form_validation->set_rules('user_address', "Address", "required|regex_match[/^[a-zA-Z]+\ +[0-9]+$/]");
-        $this->form_validation->set_rules('user_brgy', "Barangay", "required|callback__alpha_dash_space|min_length[2]");
-        $this->form_validation->set_rules('user_city', "City", "required|callback__alpha_dash_space|min_length[2]");
+        $this->form_validation->set_rules('user_address', "Address", "required");
+        $this->form_validation->set_rules("user_email", "Email", "required|valid_email");
+        $this->form_validation->set_rules("user_contact_no", "Mobile Phone No.", "required|numeric|regex_match[^(09|\+639)\d{9}$^]");
         if ($this->form_validation->run() == FALSE) {
             //ERROR IN FORM
             $this->edit_profile();
@@ -144,49 +148,18 @@ class Profile extends CI_Controller {
                 'user_sex' => $this->input->post("user_sex"),
                 'user_bday' => strtotime($this->input->post('user_bday')),
                 'user_address' => $this->input->post("user_address"),
-                'user_brgy' => $this->input->post("user_brgy"),
-                'user_city' => $this->input->post("user_city"),
+                'user_contact_no' => $this->input->post("user_contact_no"),
+                'user_email' => $this->input->post("user_email"),
                 'user_updated_at' => time()
             );
 
             if ($this->Profile_model->update_user_record($data, array("user_id" => $userDetails->user_id))) {
                 //SUCCESS
-                $this->session->set_flashdata("uploading_success", "Successfully update the record of " . $userDetails->user_lastname);
+                $this->SaveEventUser->trail($this->session->userdata("userid"), $userDetails->user_firstname . " change account information.");
+                $this->session->set_flashdata("uploading_success", "You have successfully changed your account information");
+                redirect(base_url() . "Profile/edit_profile");
             } else {
-                $this->session->set_flashdata("uploading_fail2", $userDetails->user_lastname . " seems to not exist in the database.");
-            }
-            redirect(base_url() . "Profile/");
-        }
-    }
-
-    public function edit_loginInfo_submit() {
-        $userDetails = $this->Profile_model->fetch("user", array("user_id" => $this->session->userdata("userid")))[0];
-        if (!empty($this->input->post('user_password'))) {
-
-            $data = array(
-                'user_username' => $this->input->post("user_username"),
-                'user_password' => $this->input->post("user_password"),
-                'user_updated_at' => time()
-            );
-
-            if ($this->Profile_model->update_user_record($data, array("user_id" => $userDetails->user_id))) {
-                //SUCCESS
-                $this->session->set_flashdata("uploading_success", "Successfully update the record of " . $userDetails->user_lastname);
-            } else {
-                $this->session->set_flashdata("uploading_fail2", $userDetails->user_lastname . " seems to not exist in the database.");
-            }
-            redirect(base_url() . "Profile/");
-        } else {
-            $data = array(
-                'user_username' => $this->input->post("user_username"),
-                'user_updated_at' => time()
-            );
-
-            if ($this->Profile_model->update_user_record($data, array("user_id" => $userDetails->user_id))) {
-                //SUCCESS
-                $this->session->set_flashdata("uploading_success", "Successfully update the record of " . $userDetails->user_lastname);
-            } else {
-                $this->session->set_flashdata("uploading_fail2", $userDetails->user_lastname . " seems to not exist in the database.");
+                $this->session->set_flashdata("uploading_fail", $userDetails->user_lastname . " seems to not exist in the database.");
             }
             redirect(base_url() . "Profile/");
         }
